libft v2

+++ src/put/ft_putnbr_fd.c
@@ -3,7 +3,7 @@
 /*                                                        ::::::::            */
 /*   ft_putnbr_fd.c                                     :+:    :+:            */
 /*                                                     +:+                    */
-/*   By: mweverli <marvin@codam.nl>                   +#+                     */
+/*   By: mweverli <mweverli@codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/11/08 15:27:40 by mweverli      #+#    #+#                 */
 /*   Updated: 2021/11/14 20:09:27 by mweverli      ########   odam.nl         */
@@ -16,7 +16,7 @@ void	ft_putnbr_fd(int n, int fd)
 {
 	long	b;
 
-	b = n;
+	b = (long) n;
 	if (b < 0)
 	{
 		write(fd, "-", 1);
